[rename] 존재하는 reposiroty 위치 mysql 폴더로 이동
